backend/requirements.txt
# Core Backend Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI/ML Stack
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.25.0
safetensors>=0.4.0
huggingface-hub>=0.19.0

# OpenAI for Narrative Generation
openai>=1.0.0

# Utilities
pillow>=10.0.0
numpy>=1.24.0
python-dotenv==1.0.0
